Moves methods to extract objectives and residuals to plotter since they were only used there.
