+++ fast_pedago/utils/__init__.py
@@ -1,5 +1,3 @@
 from .functions import (
     _image_from_path,
-    _extract_objective,
-    _extract_residuals,
 )
