+++ fast_pedago/utils/functions.py
@@ -1,7 +1,5 @@
 import os.path as pth
 
-import openmdao.api as om
-
 import ipywidgets as widgets
 import ipyvuetify as v
 
@@ -37,61 +35,6 @@ def _image_from_path(file_path: str, max_height: str = "52px") -> v.Html:
     )
 
     return image_widget
-
-
-def _extract_residuals(recorder_database_file_path: str) -> list:
-    """
-    From the file path to a recorder data base, extract the value of the
-    relative error of the residuals at each iteration.
-
-    :param recorder_database_file_path: absolute path to the recorder database
-    :return: two arrays containing the iterations and the associated values of
-        the relative error.
-    """
-
-    case_reader = om.CaseReader(recorder_database_file_path)
-
-    # Will only work if the recorder was attached to the base solver
-    solver_cases = case_reader.list_cases("root.nonlinear_solver")
-
-    # For the display, first iteration will be 1
-    iterations, relative_error = zip(
-        *[
-            (i + 1, case_reader.get_case(case_id).rel_err)
-            for i, case_id in enumerate(solver_cases)
-        ]
-    )
-
-    return iterations, relative_error
-
-
-def _extract_objective(recorder_database_file_path: str) -> list:
-    """
-    From the file path to a recorder data base, extract the value of the
-    objective at each iteration of the driver.
-
-    :param recorder_database_file_path: absolute path to the recorder database
-    :return: an array containing the iterations and the associated values of
-        the objective.
-    """
-
-    case_reader = om.CaseReader(recorder_database_file_path)
-
-    # Will only work if the recorder was attached to the base solver
-    solver_cases = case_reader.list_cases("driver")
-
-    # For the display, first iteration will be 1
-    iterations, objective = zip(
-        *[
-            (
-                i + 1,
-                float(list(case_reader.get_case(case_id).get_objectives().values())[0]),
-            )
-            for i, case_id in enumerate(solver_cases)
-        ]
-    )
-
-    return iterations, objective
 
 
 def _n2_xdsm_to_vue_template(html_file_path: str):
